Remove CUDA work-around and fix CUDA bugs
Some existing CUDA benchmarks have a bug -- they violate a
precondition of mdspan that the data handle needs to be an
accessible range for [0, required_span_size()), by using
nullptr as the data handle but an extents with nonzero size.

This commit implements a straightforward fix, which is to
create the layout mapping first and then use that to find
the required span size for the allocation, before finally
creating the mdspan to view the allocation.

This commit also removes a CUDA work-around for mdspan's
default constructor.